dance_agent_system/requirements.txt
google-genai
google-adk
python-dotenv
beautifulsoup4
requests
duckduckgo-search
cloudpickle
